Create report_tax.sh
An example of a scheduled script designed to produce console output (that gets emailed by cronscript.sh).
The 'if [ -d ../parms ]'  clause is useful if you want to run this script outside cronscript.sh (during development and debugging, for example). 
The 'MONTH=`date +"%b"`; if [ "${MONTH}" = "May" -o "${MONTH}" = "Dec" ]' means that although the script may be scheduled (via cronscript.sh) to run monthly, it will only produce output in May and December.
The MySQL get_uk_tax_report procedure is defined elsewhere, but for the purposes of illustration it just dumps out MySQL data in a table to console; 'mysql --html' makes sure its in html format and cronscript.sh captures this output and stores it in ${HOME}/mail to be sent to ${EMAILTO} a few minutes after it is completed.