Create update_house_prices.sh
Script to update a 'virtual' house price index.
**This will not work for you** and is provided as an example only; it only tracks GBP prices of terrace houses in Reading UK (because that's all I'm interested in).
Firstly, you need to set up a new currency in GnuCash called 'XHS' (the name is arbitrary). This will track the price of one standard house of the variety in which you are interested (Reading terraces, in my case). You can't set up new currencies through the GnuCash GUI; you have to do it by adding a row to the commodities table directly, and you need to be careful not to use a _real_ currency name, and a unique guid (see 'new_guid' function in GitHub : ohingardail/gnucash-mysql-additions )
Secondly you need to find some source of house price index data (in the UK you can get this from landregistry.data.gov.uk) and work out to extract required data from it. In my case I've tried to make sure I've got a CSV of monthly prices.
Thirdly, in order to get a reasonable estimate of _your_ property's value, you need to create a new account through the GnuCash client that uses XHS as its currency, and add a transaction for the time you bought the house for the right value (which may be 1.1 XHS or 0.9 XHS, or whatever).