Create update_rpi.sh
Script to keep a track of inflation-corrected currency value, if you want to run GnuCash net value reports using the 'real' inflation-adjusted value of your currency rather than the 'nominal' value.
This script will only work for you unamended **if** you use GBP as your base currency and the UK RPI (retail prices index) as your inflation measure, but it could easily be amended to other currencies and measures once you worked out what and where they are.
You need to pick a baseline date to start with - say 1 Jan 2000 - and dictate to GnuCash, using the client GUI prices tool - that the value of currency XXX (a 'test' currency already installed in, but unused by Gnucash) was exactly 1 GBP on that date. 
The script will use the %age values in the ONS RPI file to calculate the value of XXX in GBP for each month since that date, and update for each new month thereafter whenever it is called (or do nothing).
MySQL custom procedures 'post_commodity_price' and 'get_commodity_value' are defined in GitHub : ohingardail/gnucash-mysql-additions.